LPD-81801 Avoid CompanyThreadLocal.lock and CompanyThreadLocal.setCompanyId*
