+++ modules/util/source-formatter/source-formatter.properties
@@ -12,8 +12,6 @@
 
     checkstyle.BatchableUpdateCheck.enabled=true
 
-    checkstyle.CompanyThreadLocalCheck.enabled=true
-
     checkstyle.InstanceInitializerCheck.enabled=true
 
     checkstyle.RedundantLogCheck.enabled=true
