+++ modules/util/source-formatter/src/main/resources/checkstyle.xml
@@ -309,13 +309,6 @@
 			<message key="company.local.service.use.sql" value="Use &quot;CompanyLocalService.forEachCompany&quot; instead of SQL" />
 			<message key="portal.instances.use" value="Use &quot;PortalInstancePool.getCompanyIds&quot; instead to fetch company IDs" />
 		</module>
-		<module name="com.liferay.source.formatter.checkstyle.check.CompanyThreadLocalCheck">
-			<property name="category" value="Bug Prevention" />
-			<property name="description" value="Checks usage of `CompanyThreadLocal`." />
-			<property name="enabled" value="false" />
-			<message key="close.call.missing" value="Missing calling &quot;close&quot; to variable &quot;{0}&quot;, use &quot;{1}.close&quot; or try-with-resources statement instead" />
-			<message key="set.company.id.call.avoid" value="Do not use &quot;CompanyThreadLocal.setCompanyId&quot;, use &quot;CompanyThreadLocal.setCompanyIdWithSafeCloseable&quot; instead" />
-		</module>
 		<module name="com.liferay.source.formatter.checkstyle.check.ComponentAnnotationCheck">
 			<property name="category" value="Bug Prevention" />
 			<property name="description" value="Performs several checks on classes with @Component annotation." />
@@ -832,6 +825,11 @@
 			<property name="enabled" value="false" />
 			<message key="empty.line.missing.after.variable.definition" value="There should be an empty line after variable definition on line &quot;{0}&quot;" />
 			<message key="variable.declaration.move" value="Move the variable declaration for &quot;{0}&quot; on line &quot;{1}&quot;" />
+		</module>
+		<module name="com.liferay.source.formatter.checkstyle.check.SafeCloseableMissingCloseCheck">
+			<property name="category" value="Bug Prevention" />
+			<property name="description" value="Finds missing `close()` calls for `SafeCloseable` instances." />
+			<message key="close.call.missing" value="Missing calling &quot;close&quot; to variable &quot;{0}&quot;, use &quot;{1}.close&quot; or try-with-resources statement instead" />
 		</module>
 		<module name="com.liferay.source.formatter.checkstyle.check.SealedAndNonsealedModifierCheck">
 			<property name="category" value="Performance" />
