+++ source-formatter.properties
@@ -226,8 +226,6 @@
 
     checkstyle.CompanyIterationCheck.enabled=true
 
-    checkstyle.CompanyThreadLocalCheck.enabled=true
-
     checkstyle.ComponentExposureCheck.enabled=true
 
     checkstyle.ConstructorGlobalVariableDeclarationCheck.enabled=true
@@ -1013,6 +1011,8 @@
 
     source.check.JavaCompanyScopedIdsCheck.enabled=true
 
+    source.check.JavaCompanyThreadLocalCheck.enabled=true
+
     source.check.JavaComponentAnnotationsCheck.allowedImmediateAttributeClassNames=\
         modules/apps/document-library/document-library-service/src/main/java/com/liferay/document/library/internal/util/DLProcessorRegistryImpl.java,\
         modules/apps/static/portal-bundle-blacklist/portal-bundle-blacklist-impl/src/main/java/com/liferay/portal/bundle/blacklist/internal/BundleBlacklistManagerImpl.java
@@ -1317,8 +1317,6 @@
     source.check.JavaStaticMethodCheck.checkFinal=true
 
     source.check.JavaUniqueUpgradeProcessCheck.enabled=true
-
-    source.check.JavaUpgradeCompanyThreadLocalCheck.enabled=true
 
     source.check.JavaUpgradeCreateTableCheck.enabled=true
